pipeliner: move 'BaseParam', 'FunctionParam', 'PathExistsParam' and 'PathJoinParam' from bcl2fastq/pipeline.
