+++ auto_process_ngs/test/bcl2fastq/test_pipeline.py
@@ -13,11 +13,7 @@ from auto_process_ngs.mock import MockBcl2fastq2Exe
 from auto_process_ngs.mock import MockCellrangerExe
 from auto_process_ngs.mock import Mock10xPackageExe
 from auto_process_ngs.mock import make_mock_bcl2fastq2_output
-from auto_process_ngs.pipeliner import PipelineParam
 from auto_process_ngs.bcl2fastq.pipeline import MakeFastqs
-from auto_process_ngs.bcl2fastq.pipeline import PathJoinParam
-from auto_process_ngs.bcl2fastq.pipeline import PathExistsParam
-from auto_process_ngs.bcl2fastq.pipeline import FunctionParam
 from auto_process_ngs.bcl2fastq.pipeline import subset
 from auto_process_ngs.bcl2fastq_utils import make_custom_sample_sheet
 
@@ -4499,146 +4495,6 @@ Sample2,Sample2,,,D702,CGTGTAGG,D501,ATGTAACT,,
                               subset(lanes=[1,]),
                               subset(lanes=[2,3,4])
                           ))
-
-class TestPathJoinParam(unittest.TestCase):
-    """
-    Tests for the 'PathJoinParam' class
-    """
-    def test_pathjoinparam_with_strings(self):
-        """
-        PathJoinParam: handle static strings
-        """
-        p1 = "/mnt/data"
-        p2 = "test.txt"
-        pth = PathJoinParam(p1,p2)
-        self.assertEqual(pth.value,"/mnt/data/test.txt")
-
-    def test_pathjoinparam_with_pipelineparams(self):
-        """
-        PathJoinParam: handle PipelineParams
-        """
-        p1 = PipelineParam(value="/mnt/data")
-        p2 = PipelineParam(value="test.txt")
-        pth = PathJoinParam(p1,p2)
-        self.assertEqual(pth.value,"/mnt/data/test.txt")
-        p2.set("updated.txt")
-        self.assertEqual(pth.value,"/mnt/data/updated.txt")
-
-    def test_pathjoinparam_with_mixed_input(self):
-        """
-        PathJoinParam: handle mixture of string and PipelineParam
-        """
-        p1 = PipelineParam(value="/mnt/data")
-        p2 = "test.txt"
-        pth = PathJoinParam(p1,p2)
-        self.assertEqual(pth.value,"/mnt/data/test.txt")
-
-class TestPathExistsParam(unittest.TestCase):
-    """
-    Tests for the 'PathExistsParam' class
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestPathExistsParam')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-    
-    def test_pathexistsparam_with_strings(self):
-        """
-        PathExistsParam: handle static strings
-        """
-        dir_path = self.wd
-        file_path = os.path.join(self.wd,"file.txt")
-        with open(file_path,'wt') as fp:
-            fp.write("test\n")
-        missing_path = os.path.join(self.wd,"missing.txt")
-        self.assertTrue(PathExistsParam(dir_path).value)
-        self.assertTrue(PathExistsParam(file_path).value)
-        self.assertFalse(PathExistsParam(missing_path).value)
-
-    def test_pathexistsparam_with_pipelineparams(self):
-        """
-        PathExistsParam: handle PipelineParams
-        """
-        dir_path = PipelineParam(value=self.wd)
-        file_path = PipelineParam(value=os.path.join(self.wd,"file.txt"))
-        with open(file_path.value,'wt') as fp:
-            fp.write("test\n")
-        missing_path = PipelineParam(
-            value=os.path.join(self.wd,"missing.txt"))
-        self.assertTrue(PathExistsParam(dir_path).value)
-        self.assertTrue(PathExistsParam(file_path).value)
-        self.assertFalse(PathExistsParam(missing_path).value)
-
-    def test_pathexistsparam_with_pathjoinparams(self):
-        """
-        PathExistsParam: handle PathJoinParams
-        """
-        dir_path = PathJoinParam(self.wd)
-        file_path = PathJoinParam(self.wd,"file.txt")
-        with open(file_path.value,'wt') as fp:
-            fp.write("test\n")
-        missing_path = PathJoinParam(self.wd,"missing.txt")
-        self.assertTrue(PathExistsParam(dir_path).value)
-        self.assertTrue(PathExistsParam(file_path).value)
-        self.assertFalse(PathExistsParam(missing_path).value)
-
-class TestFunctionParam(unittest.TestCase):
-    """
-    Tests for the 'FunctionParam' class
-    """
-    #@unittest.skip("Skipped")
-    def test_functionparam_with_static_values(self):
-        """
-        FunctionParam: handle static values
-        """
-        # Lambda function
-        func_param = FunctionParam(lambda x: x,"hello")
-        self.assertEqual(func_param.value,"hello")
-        func_param = FunctionParam(lambda x,y: x + y,1,2)
-        self.assertEqual(func_param.value,3)
-        # Function with keywords
-        def func(x,y,z=None):
-            return (x,y,z)
-        func_param = FunctionParam(func,"hello","goodbye")
-        self.assertEqual(func_param.value,("hello","goodbye",None))
-        func_param = FunctionParam(func,"hello","goodbye",z="surprise!")
-        self.assertEqual(func_param.value,("hello","goodbye","surprise!"))
-
-    def test_functionparam_with_parameters(self):
-        """
-        FunctionParam: handle pipeline parameters
-        """
-        # Lambda function
-        p = PipelineParam(value="hello")
-        func_param = FunctionParam(lambda x: x,p)
-        self.assertEqual(func_param.value,"hello")
-        p.set("goodbye")
-        self.assertEqual(func_param.value,"goodbye")
-        px = PipelineParam(value=1)
-        py = PipelineParam(value=2)
-        func_param = FunctionParam(lambda x,y: x + y,px,py)
-        self.assertEqual(func_param.value,3)
-        px.set(3)
-        py.set(4)
-        self.assertEqual(func_param.value,7)
-        # Function with keywords
-        def func(x,y,z=None):
-            return (x,y,z)
-        px = PipelineParam(value="hello")
-        py = PipelineParam(value="goodbye")
-        pz = PipelineParam(value="surprise!")
-        func_param = FunctionParam(func,px,py)
-        self.assertEqual(func_param.value,("hello","goodbye",None))
-        func_param = FunctionParam(func,px,py,z=pz)
-        self.assertEqual(func_param.value,("hello","goodbye","surprise!"))
-        px.set("goodbye")
-        py.set("hello")
-        pz.set("backwards")
-        self.assertEqual(func_param.value,("goodbye","hello","backwards"))
 
 class TestSubsetFunction(unittest.TestCase):
     """
