+++ auto_process_ngs/test/test_pipeliner.py
@@ -24,8 +24,14 @@ from auto_process_ngs.pipeliner import PipelineParam
 from auto_process_ngs.pipeliner import PipelineFailure
 from auto_process_ngs.pipeliner import FileCollector
 from auto_process_ngs.pipeliner import Dispatcher
+from auto_process_ngs.pipeliner import PathJoinParam
+from auto_process_ngs.pipeliner import PathExistsParam
+from auto_process_ngs.pipeliner import FunctionParam
 from auto_process_ngs.pipeliner import resolve_parameter
 from bcftbx.JobRunner import SimpleJobRunner
+
+# Set to False to keep test output dirs
+REMOVE_TEST_OUTPUTS = True
 
 # Unit tests
 
@@ -2000,6 +2006,146 @@ class TestFileCollector(unittest.TestCase):
         self.assertEqual(len(txt_files),1)
         self.assertEqual(list(txt_files),
                          [os.path.join(self.working_dir,"test1.txt")])
+
+class TestPathJoinParam(unittest.TestCase):
+    """
+    Tests for the 'PathJoinParam' class
+    """
+    def test_pathjoinparam_with_strings(self):
+        """
+        PathJoinParam: handle static strings
+        """
+        p1 = "/mnt/data"
+        p2 = "test.txt"
+        pth = PathJoinParam(p1,p2)
+        self.assertEqual(pth.value,"/mnt/data/test.txt")
+
+    def test_pathjoinparam_with_pipelineparams(self):
+        """
+        PathJoinParam: handle PipelineParams
+        """
+        p1 = PipelineParam(value="/mnt/data")
+        p2 = PipelineParam(value="test.txt")
+        pth = PathJoinParam(p1,p2)
+        self.assertEqual(pth.value,"/mnt/data/test.txt")
+        p2.set("updated.txt")
+        self.assertEqual(pth.value,"/mnt/data/updated.txt")
+
+    def test_pathjoinparam_with_mixed_input(self):
+        """
+        PathJoinParam: handle mixture of string and PipelineParam
+        """
+        p1 = PipelineParam(value="/mnt/data")
+        p2 = "test.txt"
+        pth = PathJoinParam(p1,p2)
+        self.assertEqual(pth.value,"/mnt/data/test.txt")
+
+class TestPathExistsParam(unittest.TestCase):
+    """
+    Tests for the 'PathExistsParam' class
+    """
+    def setUp(self):
+        # Create a temp working dir
+        self.wd = tempfile.mkdtemp(suffix='TestPathExistsParam')
+
+    def tearDown(self):
+        # Remove the temporary test directory
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.wd)
+
+    def test_pathexistsparam_with_strings(self):
+        """
+        PathExistsParam: handle static strings
+        """
+        dir_path = self.wd
+        file_path = os.path.join(self.wd,"file.txt")
+        with open(file_path,'wt') as fp:
+            fp.write("test\n")
+        missing_path = os.path.join(self.wd,"missing.txt")
+        self.assertTrue(PathExistsParam(dir_path).value)
+        self.assertTrue(PathExistsParam(file_path).value)
+        self.assertFalse(PathExistsParam(missing_path).value)
+
+    def test_pathexistsparam_with_pipelineparams(self):
+        """
+        PathExistsParam: handle PipelineParams
+        """
+        dir_path = PipelineParam(value=self.wd)
+        file_path = PipelineParam(value=os.path.join(self.wd,"file.txt"))
+        with open(file_path.value,'wt') as fp:
+            fp.write("test\n")
+        missing_path = PipelineParam(
+            value=os.path.join(self.wd,"missing.txt"))
+        self.assertTrue(PathExistsParam(dir_path).value)
+        self.assertTrue(PathExistsParam(file_path).value)
+        self.assertFalse(PathExistsParam(missing_path).value)
+
+    def test_pathexistsparam_with_pathjoinparams(self):
+        """
+        PathExistsParam: handle PathJoinParams
+        """
+        dir_path = PathJoinParam(self.wd)
+        file_path = PathJoinParam(self.wd,"file.txt")
+        with open(file_path.value,'wt') as fp:
+            fp.write("test\n")
+        missing_path = PathJoinParam(self.wd,"missing.txt")
+        self.assertTrue(PathExistsParam(dir_path).value)
+        self.assertTrue(PathExistsParam(file_path).value)
+        self.assertFalse(PathExistsParam(missing_path).value)
+
+class TestFunctionParam(unittest.TestCase):
+    """
+    Tests for the 'FunctionParam' class
+    """
+    #@unittest.skip("Skipped")
+    def test_functionparam_with_static_values(self):
+        """
+        FunctionParam: handle static values
+        """
+        # Lambda function
+        func_param = FunctionParam(lambda x: x,"hello")
+        self.assertEqual(func_param.value,"hello")
+        func_param = FunctionParam(lambda x,y: x + y,1,2)
+        self.assertEqual(func_param.value,3)
+        # Function with keywords
+        def func(x,y,z=None):
+            return (x,y,z)
+        func_param = FunctionParam(func,"hello","goodbye")
+        self.assertEqual(func_param.value,("hello","goodbye",None))
+        func_param = FunctionParam(func,"hello","goodbye",z="surprise!")
+        self.assertEqual(func_param.value,("hello","goodbye","surprise!"))
+
+    def test_functionparam_with_parameters(self):
+        """
+        FunctionParam: handle pipeline parameters
+        """
+        # Lambda function
+        p = PipelineParam(value="hello")
+        func_param = FunctionParam(lambda x: x,p)
+        self.assertEqual(func_param.value,"hello")
+        p.set("goodbye")
+        self.assertEqual(func_param.value,"goodbye")
+        px = PipelineParam(value=1)
+        py = PipelineParam(value=2)
+        func_param = FunctionParam(lambda x,y: x + y,px,py)
+        self.assertEqual(func_param.value,3)
+        px.set(3)
+        py.set(4)
+        self.assertEqual(func_param.value,7)
+        # Function with keywords
+        def func(x,y,z=None):
+            return (x,y,z)
+        px = PipelineParam(value="hello")
+        py = PipelineParam(value="goodbye")
+        pz = PipelineParam(value="surprise!")
+        func_param = FunctionParam(func,px,py)
+        self.assertEqual(func_param.value,("hello","goodbye",None))
+        func_param = FunctionParam(func,px,py,z=pz)
+        self.assertEqual(func_param.value,("hello","goodbye","surprise!"))
+        px.set("goodbye")
+        py.set("hello")
+        pz.set("backwards")
+        self.assertEqual(func_param.value,("goodbye","hello","backwards"))
 
 class TestDispatcher(unittest.TestCase):
